linebreaks and spaces fixed

+++ app/models/cart.rb
@@ -1,11 +1,13 @@
+# frozen_string_literal: true
+
 class Cart < ApplicationRecord
   has_many :cart_items, dependent: :destroy
   has_one :order, dependent: :destroy
-  
+
   belongs_to :user
- 
+
   def total
-    cart_items.collect {|item| item.valid? ? item.subtotal : 0 }.sum
+    cart_items.collect { |item| item.valid? ? item.subtotal : 0 }.sum
   end
 
   def placed_cart
@@ -13,8 +15,6 @@ class Cart < ApplicationRecord
   end
 
   def require_same_restaurant
-    unless cart_items.first.food_item.restaurant_id == cart_items.last.food_item.restaurant_id
-      cart_items.delete_all
-    end
+    self.cart_items.delete_all unless self.cart_items.first.food_item.restaurant_id == self.cart_items.last.food_item.restaurant_id
   end
 end
